modify: moved status Tests
