+++ test/managers/TaskManagerTest.java
@@ -3,6 +3,7 @@ package managers;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tasks.Epic;
+import tasks.Status;
 import tasks.Subtask;
 import tasks.Task;
 
@@ -305,5 +306,111 @@ abstract class TaskManagerTest<T extends TaskManager> {
         List<Task> prioritized = new ArrayList<>(manager.getPrioritizedTasks());
         assertEquals(subtask2, prioritized.get(0), "Ранее запланированная подзадача должна быть первой");
         assertEquals(subtask1, prioritized.get(1), "Позже запланированная подзадача должна идти следом");
+    }
+
+    @Test
+    void testEpicStatusAllSubtasksNew() {
+        Epic epic = new Epic("Epic 1", "Epic description");
+        manager.createEpic(epic);
+
+        Subtask subtask1 = new Subtask("Subtask 1", "Description", epic.getId());
+        Subtask subtask2 = new Subtask("Subtask 2", "Description", epic.getId());
+
+        manager.createSubtask(subtask1);
+        manager.createSubtask(subtask2);
+
+        assertEquals(Status.NEW, manager.getEpicById(epic.getId()).getStatus(),
+                "Статус эпика должен быть NEW, если все подзадачи NEW");
+    }
+
+    @Test
+    void testEpicStatusAllSubtasksDone() {
+        Epic epic = new Epic("Epic 1", "Epic description");
+        manager.createEpic(epic);
+
+        Subtask subtask1 = new Subtask("Subtask 1", "Description", epic.getId());
+        Subtask subtask2 = new Subtask("Subtask 2", "Description", epic.getId());
+
+        subtask1.setStatus(Status.DONE);
+        subtask2.setStatus(Status.DONE);
+
+        manager.createSubtask(subtask1);
+        manager.createSubtask(subtask2);
+        manager.updateSubtask(subtask1);
+        manager.updateSubtask(subtask2);
+
+        assertEquals(Status.DONE, manager.getEpicById(epic.getId()).getStatus(),
+                "Статус эпика должен быть DONE, если все подзадачи DONE");
+    }
+
+    @Test
+    void testEpicStatusMixedNewAndDoneSubtasks() {
+        Epic epic = new Epic("Epic 1", "Epic description");
+        manager.createEpic(epic);
+
+        Subtask subtask1 = new Subtask("Subtask 1", "Description", epic.getId());
+        Subtask subtask2 = new Subtask("Subtask 2", "Description", epic.getId());
+
+        subtask1.setStatus(Status.NEW);
+        subtask2.setStatus(Status.DONE);
+
+        manager.createSubtask(subtask1);
+        manager.createSubtask(subtask2);
+        manager.updateSubtask(subtask1);
+        manager.updateSubtask(subtask2);
+
+        assertEquals(Status.IN_PROGRESS, manager.getEpicById(epic.getId()).getStatus(),
+                "Статус эпика должен быть IN_PROGRESS, если есть подзадачи NEW и DONE");
+    }
+
+    @Test
+    void testEpicStatusAllSubtasksInProgress() {
+        Epic epic = new Epic("Epic 1", "Epic description");
+        manager.createEpic(epic);
+
+        Subtask subtask1 = new Subtask("Subtask 1", "Description", epic.getId());
+        Subtask subtask2 = new Subtask("Subtask 2", "Description", epic.getId());
+
+        subtask1.setStatus(Status.IN_PROGRESS);
+        subtask2.setStatus(Status.IN_PROGRESS);
+
+        manager.createSubtask(subtask1);
+        manager.createSubtask(subtask2);
+        manager.updateSubtask(subtask1);
+        manager.updateSubtask(subtask2);
+
+        assertEquals(Status.IN_PROGRESS, manager.getEpicById(epic.getId()).getStatus(),
+                "Статус эпика должен быть IN_PROGRESS, если все подзадачи IN_PROGRESS");
+    }
+
+    @Test
+    void testEpicStatusUpdatesWhenSubtaskStatusChanges() {
+        Epic epic = new Epic("Epic 1", "Epic description");
+        manager.createEpic(epic);
+
+        Subtask subtask = new Subtask("Subtask 1", "Description", epic.getId());
+        manager.createSubtask(subtask);
+
+        assertEquals(Status.NEW, manager.getEpicById(epic.getId()).getStatus(),
+                "Ошибка: эпик должен быть в статусе NEW.");
+
+        subtask.setStatus(Status.IN_PROGRESS);
+        manager.updateSubtask(subtask);
+        assertEquals(Status.IN_PROGRESS, manager.getEpicById(epic.getId()).getStatus(),
+                "Ошибка: эпик должен обновиться в IN_PROGRESS.");
+
+        subtask.setStatus(Status.DONE);
+        manager.updateSubtask(subtask);
+        assertEquals(Status.DONE, manager.getEpicById(epic.getId()).getStatus(),
+                "Ошибка: эпик должен обновиться в DONE.");
+    }
+
+    @Test
+    void testEpicStatusWithoutSubtasks() {
+        Epic epic = new Epic("Epic 1", "Epic description");
+        manager.createEpic(epic);
+
+        assertEquals(Status.NEW, manager.getEpicById(epic.getId()).getStatus(),
+                "Статус эпика должен быть NEW, если нет подзадач");
     }
 }
